pyproject.toml
[project]
name = "portfolio-optimization"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Onishi",email = "gabriel.onishi@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "ruff (>=0.11.13,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'